Edits
-moved methods to utils
-remove mounted
-remove pnpm-lock.yaml
